pkg/container: take mount configuration
