+++ server/container_create.go
@@ -3,180 +3,19 @@ package server
 import (
 	"context"
 	"fmt"
-	"os"
-	"path/filepath"
 	"strconv"
 	"strings"
 
-	"github.com/containers/storage/pkg/idtools"
-	"github.com/containers/storage/pkg/mount"
-	"github.com/containers/storage/pkg/stringid"
 	"github.com/cri-o/cri-o/internal/lib/sandbox"
 	"github.com/cri-o/cri-o/internal/log"
-	"github.com/cri-o/cri-o/internal/storage"
-	"github.com/cri-o/cri-o/pkg/config"
 	"github.com/cri-o/cri-o/pkg/container"
 	"github.com/cri-o/cri-o/server/cri/types"
 	"github.com/cri-o/cri-o/utils"
-	securejoin "github.com/cyphar/filepath-securejoin"
 	v1 "github.com/opencontainers/image-spec/specs-go/v1"
 	rspec "github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/opencontainers/runtime-tools/generate"
 	"github.com/pkg/errors"
 )
-
-type orderedMounts []rspec.Mount
-
-// Len returns the number of mounts. Used in sorting.
-func (m orderedMounts) Len() int {
-	return len(m)
-}
-
-// Less returns true if the number of parts (a/b/c would be 3 parts) in the
-// mount indexed by parameter 1 is less than that of the mount indexed by
-// parameter 2. Used in sorting.
-func (m orderedMounts) Less(i, j int) bool {
-	return m.parts(i) < m.parts(j)
-}
-
-// Swap swaps two items in an array of mounts. Used in sorting
-func (m orderedMounts) Swap(i, j int) {
-	m[i], m[j] = m[j], m[i]
-}
-
-// parts returns the number of parts in the destination of a mount. Used in sorting.
-func (m orderedMounts) parts(i int) int {
-	return strings.Count(filepath.Clean(m[i].Destination), string(os.PathSeparator))
-}
-
-// mounts defines how to sort runtime.Mount.
-// This is the same with the Docker implementation:
-//   https://github.com/moby/moby/blob/17.05.x/daemon/volumes.go#L26
-type criOrderedMounts []*types.Mount
-
-// Len returns the number of mounts. Used in sorting.
-func (m criOrderedMounts) Len() int {
-	return len(m)
-}
-
-// Less returns true if the number of parts (a/b/c would be 3 parts) in the
-// mount indexed by parameter 1 is less than that of the mount indexed by
-// parameter 2. Used in sorting.
-func (m criOrderedMounts) Less(i, j int) bool {
-	return m.parts(i) < m.parts(j)
-}
-
-// Swap swaps two items in an array of mounts. Used in sorting
-func (m criOrderedMounts) Swap(i, j int) {
-	m[i], m[j] = m[j], m[i]
-}
-
-// parts returns the number of parts in the destination of a mount. Used in sorting.
-func (m criOrderedMounts) parts(i int) int {
-	return strings.Count(filepath.Clean(m[i].ContainerPath), string(os.PathSeparator))
-}
-
-// Ensure mount point on which path is mounted, is shared.
-func ensureShared(path string, mountInfos []*mount.Info) error {
-	sourceMount, optionalOpts, err := getSourceMount(path, mountInfos)
-	if err != nil {
-		return err
-	}
-
-	// Make sure source mount point is shared.
-	optsSplit := strings.Split(optionalOpts, " ")
-	for _, opt := range optsSplit {
-		if strings.HasPrefix(opt, "shared:") {
-			return nil
-		}
-	}
-
-	return fmt.Errorf("path %q is mounted on %q but it is not a shared mount", path, sourceMount)
-}
-
-// Ensure mount point on which path is mounted, is either shared or slave.
-func ensureSharedOrSlave(path string, mountInfos []*mount.Info) error {
-	sourceMount, optionalOpts, err := getSourceMount(path, mountInfos)
-	if err != nil {
-		return err
-	}
-	// Make sure source mount point is shared.
-	optsSplit := strings.Split(optionalOpts, " ")
-	for _, opt := range optsSplit {
-		if strings.HasPrefix(opt, "shared:") {
-			return nil
-		} else if strings.HasPrefix(opt, "master:") {
-			return nil
-		}
-	}
-	return fmt.Errorf("path %q is mounted on %q but it is not a shared or slave mount", path, sourceMount)
-}
-
-func addImageVolumes(ctx context.Context, rootfs string, s *Server, containerInfo *storage.ContainerInfo, mountLabel string, specgen *generate.Generator) ([]rspec.Mount, error) {
-	mounts := []rspec.Mount{}
-	for dest := range containerInfo.Config.Config.Volumes {
-		fp, err := securejoin.SecureJoin(rootfs, dest)
-		if err != nil {
-			return nil, err
-		}
-		switch s.config.ImageVolumes {
-		case config.ImageVolumesMkdir:
-			IDs := idtools.IDPair{UID: int(specgen.Config.Process.User.UID), GID: int(specgen.Config.Process.User.GID)}
-			if err1 := idtools.MkdirAllAndChownNew(fp, 0o755, IDs); err1 != nil {
-				return nil, err1
-			}
-			if mountLabel != "" {
-				if err1 := securityLabel(fp, mountLabel, true); err1 != nil {
-					return nil, err1
-				}
-			}
-		case config.ImageVolumesBind:
-			volumeDirName := stringid.GenerateNonCryptoID()
-			src := filepath.Join(containerInfo.RunDir, "mounts", volumeDirName)
-			if err1 := os.MkdirAll(src, 0o755); err1 != nil {
-				return nil, err1
-			}
-			// Label the source with the sandbox selinux mount label
-			if mountLabel != "" {
-				if err1 := securityLabel(src, mountLabel, true); err1 != nil {
-					return nil, err1
-				}
-			}
-
-			log.Debugf(ctx, "Adding bind mounted volume: %s to %s", src, dest)
-			mounts = append(mounts, rspec.Mount{
-				Source:      src,
-				Destination: dest,
-				Type:        "bind",
-				Options:     []string{"private", "bind", "rw"},
-			})
-
-		case config.ImageVolumesIgnore:
-			log.Debugf(ctx, "ignoring volume %v", dest)
-		default:
-			log.Errorf(ctx, "unrecognized image volumes setting")
-		}
-	}
-	return mounts, nil
-}
-
-// resolveSymbolicLink resolves a possible symlink path. If the path is a symlink, returns resolved
-// path; if not, returns the original path.
-// note: strictly SecureJoin is not sufficient, as it does not error when a part of the path doesn't exist
-// but simply moves on. If the last part of the path doesn't exist, it may need to be created.
-func resolveSymbolicLink(scope, path string) (string, error) {
-	info, err := os.Lstat(path)
-	if err != nil {
-		return "", err
-	}
-	if info.Mode()&os.ModeSymlink != os.ModeSymlink {
-		return path, nil
-	}
-	if scope == "" {
-		scope = "/"
-	}
-	return securejoin.SecureJoin(scope, path)
-}
 
 // setupContainerUser sets the UID, GID and supplemental groups in OCI runtime config
 func setupContainerUser(ctx context.Context, specgen *generate.Generator, rootfs, mountLabel, ctrRunDir string, sc *types.LinuxContainerSecurityContext, imageConfig *v1.Image) error {
@@ -230,7 +69,7 @@ func setupContainerUser(ctx context.Context, specgen *generate.Generator, rootfs
 			return err
 		}
 		if passwdPath != "" {
-			if err := securityLabel(passwdPath, mountLabel, false); err != nil {
+			if err := container.SecurityLabel(passwdPath, mountLabel, false); err != nil {
 				return err
 			}
 
@@ -382,15 +221,6 @@ func setupCapabilities(specgen *generate.Generator, capabilities *types.Capabili
 	return nil
 }
 
-func hostNetwork(containerConfig *types.ContainerConfig) bool {
-	securityContext := containerConfig.Linux.SecurityContext
-	if securityContext == nil || securityContext.NamespaceOptions == nil {
-		return false
-	}
-
-	return securityContext.NamespaceOptions.Network == types.NamespaceModeNODE
-}
-
 // CreateContainer creates a new container in specified PodSandbox
 func (s *Server) CreateContainer(ctx context.Context, req *types.CreateContainerRequest) (res *types.CreateContainerResponse, retErr error) {
 	log.Infof(ctx, "Creating container: %s", translateLabelsToDescription(req.Config.Labels))
@@ -512,13 +342,4 @@ func (s *Server) CreateContainer(ctx context.Context, req *types.CreateContainer
 	return &types.CreateContainerResponse{
 		ContainerID: ctr.ID(),
 	}, nil
-}
-
-func isInCRIMounts(dst string, mounts []*types.Mount) bool {
-	for _, m := range mounts {
-		if m.ContainerPath == dst {
-			return true
-		}
-	}
-	return false
 }
